Presage TView implementation
The implementation will most likely involve the addition of a either a
transpose field to Dense or a Transpose field to blas matrix types.
The latter is probably the more sensible approach, though I'm not sure
at this stage since we don't actually use the blas matrix fields
directly (the client must be explicit to retain the similarity with
the cblas interface - should this change?).